# api/swagger/swager.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'BookStore API'
  description: 'An REST API of book store. Allows you to CRUD books, authors, genres, publishers. Also you can create an account and make orders'
  contact:
    name: Arsen
    email: arsenulykbekov9@gmail.com
servers:
  - url: "http://localhost:8080/"
paths:
...
  /publishers:
    description: Publisher Resource
    get:
      tags:
        - publishers
      parameters:
        - name: page
          in: query
          description: The page which you wanna get
          schema:
            type: integer
        - name: size
          in: query
          description: The amount of elements in one page
          schema:
            type: integer
      responses:
        200:
          description: Successfully returned Publisher
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/publisherDTO'
    post:
      tags:
        - publishers
      description: Create a new Publisher
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/savePublisherDTO'
      responses:
        200:
          description: Publisher successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/publisherDTO'
  /publishers/{id}:
    description: Get, Delete Publisher based on ID path parameter
    get:
      tags:
        - publishers
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Successfully returned Publisher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/publisherDTO'
        400:
          description: Invalid ID supplied
        404:
          description: Publisher not found
    put:
      tags:
        - publishers
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      description: Update an existing Publisher By Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/savePublisherDTO'
      responses:
        200:
          description: Successfully updated Publisher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/publisherDTO'

    delete:
      tags:
        - publishers
      description: Delete an Publisher by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Successfully deleted Publisher
        400:
          description: Invalid ID supplied
        404:
          description: Publisher not found
  /publishers/name:
    description: Get Publisher based on name containing
    get:
      tags:
        - publishers
      parameters:
        - in: query
          name: name
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successfully returned Publisher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/publisherDTO'
        400:
          description: Invalid ID supplied
        404:
          description: Publisher not found


  # -------------------------------------------------------------------------------------
  # -----------------------------------Genres--------------------------------------------
...
  /users:
    description: User Resource
    get:
      tags:
        - users
      responses:
        200:
          description: return a current registered user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userDTO'
    post:
      tags:
        - users
      description: Create a new user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/saveUserDTO'
      responses:
        200:
          description: Genre successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userDTO'
  /users/{id}:
    description: Get, Delete User based on ID path parameter
    get:
      tags:
        - users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: user wirh {id}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userDTO'
        403:
          description: forbiden, only users with role Admin can get access for information about another user
    delete:
      tags:
        - users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: user succesfully deleted
        403:
          description: only admim can delete others users
  /users/name:
    description: Get User based on ID path parameter
    get:
      tags:
        - users
      parameters:
        - in: query
          name: name
          required: true
          schema:
            type: string
      responses:
        200:
          description: user wirh name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userDTO'
        404:
          description: User not found
  /users/admin/users:
    description: Get All Users
    get:
      tags:
        - users
      responses:
        200:
          description: all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/userDTO'
        403:
          description: Only admin have acces for all users
  /users/update/username:
    description: update current user's username
    put:
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/saveUserDTO"
      responses:
        200:
          description: username successfully updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/userDTO'
        400:
          description: username already registered
  /users/update/password:
    description: update current user's password
    put:
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/saveUserDTO"
      responses:
        200:
          description: password successfully updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/userDTO'
  /users/update/admin/role/{id}:
    description: Admin update user's role
    put:
      tags:
        - users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/saveUserDTO"
      responses:
        200:
          description: password successfully updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/userDTO'


  # -------------------------------------------------------------------------------------
  # -----------------------------------Orders--------------------------------------------
...
components:
  schemas:
...
    bookDTO:
      type: object
      properties:
        id:
          type: integer
          example: 0
        price:
          type: number
          format: double
          example: 4000.0
        authorList:
          type: array
          items:
            type: string
            example: ['Fyodor Dostoevsky', 'Hovard Lovecraft']
        publisher:
          type: string
          example: 'ACT'
        name:
          type: string
          example: 'Clean Code'
        numOfpages:
          type: integer
          example: 300
        yearOfIssue:
          type: integer
          example: 1992
        genres:
          type: array
          items:
            type: string
            example: ['mystic', 'thriller']
        bookQuantity:
          type: integer
          example: 2
        bookImage:
          type: string
          example: "https://simg.marwin.kz/media/catalog/product/cache/41deb699a7fea062a8915debbbb0442c/c/o/akunin_b_doroga_v_kitej.jpg"
...
    publisherDTO:
      type: object
      properties:
        id:
          type: integer
          example: 0
        name:
          type: string
          example: ACT
        listOfBook:
          type: array
          items:
            $ref: '#/components/schemas/bookDTO'
    savePublisherDTO:
      type: object
      properties:
        name:
          type: string
          example: ACT

    # -------------------------------------------------------------------------------------
    #-----------------------------------Genres---------------------------------------------
...
    userDTO:
      type: object
      properties:
        id:
          type: integer
          example: 0
        username:
          type: string
          example: admin
        password:
          type: string
          example: $2a$10$87/Elk7ApmPm8wusEAqssOq5NHh0MwO6SnjrmJYbRd44vDb4VfeAq
        role:
          type: string
          example: USER
        isBlocked:
          type: boolean
          example: false
    saveUserDTO:
      type: object
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: mypassword
        role:
          type: string
          example: USER
        blocked:
          type: boolean
          example: false

    # -------------------------------------------------------------------------------------
    # ------------------------------------Orders------------------------------------------
...